Fix infinite loop, item use crash

+++ src/main/java/shukaro/warptheory/handlers/IWarpEvent.java
@@ -9,7 +9,7 @@ public abstract class IWarpEvent
 
     public abstract int getSeverity();
 
-    public final int getCost() { return (int)Math.ceil(getSeverity() / 10); }
+    public final int getCost() { return (int)Math.ceil(getSeverity() / (double)10); }
 
     public abstract boolean doEvent(World world, EntityPlayer player);
 }
